split connectors between openbis and db
