+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/DependencyManager.groovy
@@ -18,6 +18,10 @@ import life.qbic.portal.offermanager.components.person.update.UpdatePersonViewMo
 import life.qbic.portal.offermanager.dataresources.persons.AffiliationResourcesService
 import life.qbic.portal.offermanager.dataresources.persons.PersonDbConnector
 import life.qbic.portal.offermanager.dataresources.persons.CustomerResourceService
+import life.qbic.portal.offermanager.dataresources.projects.ProjectMainConnector
+import life.qbic.portal.offermanager.dataresources.projects.ProjectDbConnector
+
+import life.qbic.openbis.openbisclient.OpenBisClient
 
 import life.qbic.portal.offermanager.dataresources.database.DatabaseSession
 import life.qbic.portal.offermanager.dataresources.offers.OfferDbConnector
@@ -90,6 +94,9 @@ class DependencyManager {
     private PersonDbConnector customerDbConnector
     private OfferDbConnector offerDbConnector
     private ProductsDbConnector productsDbConnector
+    private ProjectMainConnector projectMainConnector
+    private ProjectDbConnector projectDbConnector
+    private OpenBisClient openbisClient
 
     private CreatePerson createCustomer
     private CreatePerson createCustomerNewOffer
@@ -168,6 +175,14 @@ class DependencyManager {
             productsDbConnector = new ProductsDbConnector(DatabaseSession.getInstance())
             offerDbConnector = new OfferDbConnector(DatabaseSession.getInstance(),
                     customerDbConnector, productsDbConnector)
+            projectDbConnector = new ProjectDbConnector(DatabaseSession.getInstance(), customerDbConnector)
+            
+            
+            final String openbisURL = configurationManager.getDataSourceUrl() + "/openbis/openbis";
+            openbisClient = new OpenBisClient(configurationManager.getDataSourceUser(), configurationManager.getDataSourcePassword(), openbisURL)
+            openbisClient.login();
+            
+            projectMainConnector = new ProjectMainConnector(projectDbConnector, openbisClient)
 
         } catch (Exception e) {
             log.error("Unexpected exception during customer database connection.", e)
