+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/dataresources/projects/ProjectDbConnector.groovy
@@ -1,0 +1,162 @@
+package life.qbic.portal.offermanager.dataresources.projects
+
+import groovy.util.logging.Log4j2
+import life.qbic.datamodel.dtos.business.Customer
+import life.qbic.datamodel.dtos.business.ProjectManager
+import life.qbic.datamodel.dtos.general.Person
+import life.qbic.portal.offermanager.dataresources.persons.PersonDbConnector
+
+import life.qbic.datamodel.dtos.projectmanagement.*
+
+import life.qbic.business.exceptions.DatabaseQueryException
+import life.qbic.portal.offermanager.dataresources.database.ConnectionProvider
+
+import java.sql.Connection
+import java.sql.PreparedStatement
+import java.sql.ResultSet
+import java.sql.Statement
+
+/**
+ * Provides operations on QBiC project data
+ *
+ * This class is responsible for transferring project data to the customer db
+ *
+ * @since: 1.0.0
+ *
+ */
+@Log4j2
+class ProjectDbConnector {
+
+  /**
+   * A connection to the customer database used to create queries.
+   */
+  private final ConnectionProvider connectionProvider
+  private final PersonDbConnector personDbConnector
+
+  /**
+   * Constructor for a ProjectDbConnector
+   * @param connectionProvider a connection provider to the customer db
+   * @param personDbConnector db connector used to connect projects to customer and manager
+   * 
+   */
+  ProjectDbConnector(ConnectionProvider connectionProvider, PersonDbConnector personDbConnector) {
+    this.connectionProvider = connectionProvider
+    this.personDbConnector = personDbConnector
+  }
+  
+  public Project addProjectAndConnectPersonsInUserDB(projectIdentifier, projectApplication) {
+    //collect infos needed for database
+    String projectTitle = projectApplication.getProjectTitle()
+    Customer customer = projectApplication.getCustomer()
+    ProjectManager projectManager = projectApplication.getProjectManager()
+    
+    //fetch needed person ids from database
+    int customerID = personDBConnector.getPersonId(customer)
+    int managerID = personDBConnector.getPersonId(projectManager)
+  
+    Connection connection = connectionProvider.connect()
+    connection.setAutoCommit(false)
+
+    connection.withCloseable {it ->
+      try {
+        int projectID = addProjectToDB(it, projectIdentifier, projectTitle)
+        addPersonToProject(it, projectID, managerID, "Manager")
+        addPersonToProject(it, projectID, customerID, "PI")
+      
+        it.commit()
+        
+      } catch (Exception e) {
+        log.error(e.message)
+        log.error(e.stackTrace.join("\n"))
+        it.rollback()
+        it.close() //is this needed?
+        throw new DatabaseQueryException("Could not add person and project data to user database.")
+      }
+    }
+  return new Project(projectIdentifier, projectTitle, projectApplication.getLinkedOffer())
+ }
+
+  private int isProjectInDB(String projectIdentifier) {
+    log.info("Looking for project " + projectIdentifier + " in the DB");
+    String sql = "SELECT * from projects WHERE openbis_project_identifier = ?";
+    int res = -1;
+    Connection connection = connectionProvider.connect()
+    connection.withCloseable { it ->
+      PreparedStatement statement = it.prepareStatement(sql);
+      statement.setString(1, projectIdentifier);
+      ResultSet rs = statement.executeQuery();
+      if (rs.next()) {
+        res = rs.getInt("id");
+      }
+    }
+    return res;
+  }
+
+  public int addProjectToDB(Connection connection, String projectIdentifier, String projectName) {
+    int exists = isProjectInDB(projectIdentifier);
+    if (exists < 0) {
+      log.info("Trying to add project " + projectIdentifier + " to the person DB");
+      String sql = "INSERT INTO projects (openbis_project_identifier, short_title) VALUES(?, ?)";
+      try (PreparedStatement statement =
+      connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+        statement.setString(1, projectIdentifier);
+        statement.setString(2, projectName);
+        statement.execute();
+        ResultSet rs = statement.getGeneratedKeys();
+        if (rs.next()) {
+          logout(conn);
+          log.info("Successful.");
+          return rs.getInt(1);
+        }
+      } catch (Exception e) {
+        log.error("SQL operation unsuccessful: " + e.getMessage());
+        e.printStackTrace();
+      }
+      return -1;
+    }
+    return exists;
+  }
+
+  public void addPersonToProject(Connection connection, int projectID, int personID, String role) {
+    if (!hasPersonRoleInProject(personID, projectID, role)) {
+      log.info("Trying to add person with role " + role + " to a project.");
+      String sql =
+      "INSERT INTO projects_persons (project_id, person_id, project_role) VALUES(?, ?, ?)";
+      try (PreparedStatement statement =
+      connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+        statement.setInt(1, projectID);
+        statement.setInt(2, personID);
+        statement.setString(3, role);
+        statement.execute();
+        log.info("Successful.");
+      } catch (Exception e) {
+        log.error("SQL operation unsuccessful: " + e.getMessage());
+        e.printStackTrace();
+      }
+    }
+  }
+
+  public boolean hasPersonRoleInProject(int personID, int projectID, String role) {
+    logger.info("Checking if person already has this role in the project.");
+    String sql =
+        "SELECT * from projects_persons WHERE person_id = ? AND project_id = ? and project_role = ?";
+    boolean res = false;
+  Connection connection = connectionProvider.connect()
+    try {
+      PreparedStatement statement = connection.prepareStatement(sql);
+      statement.setInt(1, personID);
+      statement.setInt(2, projectID);
+      statement.setString(3, role);
+      ResultSet rs = statement.executeQuery();
+      if (rs.next()) {
+        res = true;
+        logger.info("person already has this role!");
+      }
+    } catch (Exception e) {
+      logger.error("SQL operation unsuccessful: " + e.getMessage());
+      e.printStackTrace();
+    }
+    logout(conn);
+    return res;
+  }
+}
